@@ -1,11 +1,7 @@
-import Vue from 'vue'
-import App from './App.vue'
-import './registerServiceWorker'
-import router from './router'
+import { createApp } from "vue";
+import App from "./App.vue";
+import router from "./router";
 
-Vue.config.productionTip = false
-
-new Vue({
-	router,
-	render: h => h(App),
-}).$mount('#app')
+const app = createApp(App);
+app.use(router);
+app.mount("#app");
