@@ -1,34 +1,32 @@
-import Vue from 'vue'
-import VueRouter from 'vue-router'
-import Home from '../views/Home.vue'
-
-import Portfolio from '../views/Portfolio.vue'
-
-Vue.use(VueRouter)
+import { createRouter, createWebHistory } from "vue-router";
+import HomePage from "../views/Home.vue";
+import PortfolioPage from "../views/Portfolio.vue";
 
 const routes = [
-	{
-		path: '/',
-		name: 'home',
-		component: Home,
-	},
-	{
-		path: '/about',
-		name: 'about',
-		// route level code-splitting
-		// this generates a separate chunk (about.[hash].js) for this route
-		// which is lazy-loaded when the route is visited.
-		component: () => import(/* webpackChunkName: "about" */ '../views/About.vue'),
-	},
-	{
-		path: '/portfolio',
-		name: 'portfolio',
-		component: Portfolio,
-	},
-]
+  {
+    path: "/",
+    name: "home",
+    component: HomePage,
+  },
+  {
+    path: "/about",
+    name: "about",
+    // route level code-splitting
+    // this generates a separate chunk (about.[hash].js) for this route
+    // which is lazy-loaded when the route is visited.
+    component: () =>
+      import(/* webpackChunkName: "about" */ "../views/About.vue"),
+  },
+  {
+    path: "/portfolio",
+    name: "portfolio",
+    component: PortfolioPage,
+  },
+];
 
-const router = new VueRouter({
-	routes,
-})
+const router = createRouter({
+  history: createWebHistory(process.env.BASE_URL),
+  routes,
+});
 
-export default router
+export default router;
