@@ -1,4 +1,8 @@
+/* eslint-disable */
 declare module '*.vue' {
-  import Vue from 'vue'
-  export default Vue
+  import type { DefineComponent } from 'vue'
+  const component: DefineComponent<{}, {}, any>
+  export default component
 }
+
+declare module 'vue-router'
